+++ src/main/java/com/epam/reportportal/service/LaunchLoggingContext.java
@@ -19,13 +19,11 @@ package com.epam.reportportal.service;
 import com.epam.reportportal.listeners.ListenerParameters;
 import com.epam.reportportal.message.TypeAwareByteSource;
 import com.epam.reportportal.service.logs.LogBatchingFlowable;
+import com.epam.reportportal.service.logs.LoggingSubscriber;
 import com.epam.reportportal.utils.http.HttpRequestUtils;
 import com.epam.ta.reportportal.ws.model.BatchSaveOperatingRS;
 import com.epam.ta.reportportal.ws.model.log.SaveLogRQ;
-import io.reactivex.Completable;
-import io.reactivex.Flowable;
-import io.reactivex.Maybe;
-import io.reactivex.Scheduler;
+import io.reactivex.*;
 import io.reactivex.functions.Function;
 import io.reactivex.internal.operators.flowable.FlowableFromObservable;
 import io.reactivex.plugins.RxJavaPlugins;
@@ -36,7 +34,6 @@ import javax.annotation.Nullable;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static com.epam.reportportal.utils.SubscriptionUtils.logFlowableResults;
 import static com.epam.reportportal.utils.files.ImageConverter.convert;
 import static com.epam.reportportal.utils.files.ImageConverter.isImage;
 import static com.google.common.io.ByteSource.wrap;
@@ -65,8 +62,8 @@ public class LaunchLoggingContext {
 	/* Whether Image should be converted to BlackAndWhite */
 	private final boolean convertImages;
 
-	private LaunchLoggingContext(Maybe<String> launchUuid, final ReportPortalClient client, Scheduler scheduler,
-			ListenerParameters parameters) {
+	private LaunchLoggingContext(Maybe<String> launchUuid, ReportPortalClient client, Scheduler scheduler, ListenerParameters parameters,
+			FlowableSubscriber<BatchSaveOperatingRS> loggingSubscriber) {
 		this.launchUuid = launchUuid;
 		this.emitter = PublishSubject.create();
 		this.convertImages = parameters.isConvertImage();
@@ -76,15 +73,31 @@ public class LaunchLoggingContext {
 				))
 				.flatMap((Function<List<SaveLogRQ>, Flowable<BatchSaveOperatingRS>>) rqs -> client.log(HttpRequestUtils.buildLogMultiPartRequest(
 						rqs)).toFlowable())
-				.doOnError(Throwable::printStackTrace)
 				.observeOn(scheduler)
 				.onBackpressureBuffer(parameters.getRxBufferSize(), false, true)
-				.subscribe(logFlowableResults("Launch logging context"));
+				.subscribe(loggingSubscriber);
 	}
 
 	@Nullable
 	public static LaunchLoggingContext context(String key) {
 		return loggingContextMap.get(key);
+	}
+
+	/**
+	 * Initializes new logging context and attaches it to current thread
+	 *
+	 * @param launchUuid        a UUID of a Launch
+	 * @param client            Client of ReportPortal
+	 * @param scheduler         a {@link Scheduler} to use with this LoggingContext
+	 * @param parameters        Report Portal client configuration parameters
+	 * @param loggingSubscriber RxJava subscriber on logging results
+	 * @return New Logging Context
+	 */
+	public static LaunchLoggingContext init(Maybe<String> launchUuid, final ReportPortalClient client, Scheduler scheduler,
+			ListenerParameters parameters, FlowableSubscriber<BatchSaveOperatingRS> loggingSubscriber) {
+		LaunchLoggingContext context = new LaunchLoggingContext(launchUuid, client, scheduler, parameters, loggingSubscriber);
+		loggingContextMap.put(DEFAULT_LAUNCH_KEY, context);
+		return context;
 	}
 
 	/**
@@ -98,9 +111,7 @@ public class LaunchLoggingContext {
 	 */
 	public static LaunchLoggingContext init(Maybe<String> launchUuid, final ReportPortalClient client, Scheduler scheduler,
 			ListenerParameters parameters) {
-		LaunchLoggingContext context = new LaunchLoggingContext(launchUuid, client, scheduler, parameters);
-		loggingContextMap.put(DEFAULT_LAUNCH_KEY, context);
-		return context;
+		return init(launchUuid, client, scheduler, parameters, new LoggingSubscriber());
 	}
 
 	/**
@@ -130,9 +141,7 @@ public class LaunchLoggingContext {
 		ListenerParameters params = new ListenerParameters();
 		params.setBatchLogsSize(batchLogsSize);
 		params.setConvertImage(convertImages);
-		LaunchLoggingContext context = new LaunchLoggingContext(launchUuid, client, scheduler, params);
-		loggingContextMap.put(DEFAULT_LAUNCH_KEY, context);
-		return context;
+		return init(launchUuid, client, scheduler, params, new LoggingSubscriber());
 	}
 
 	/**