+++ src/main/java/com/epam/reportportal/service/LoggingContext.java
@@ -18,14 +18,12 @@ package com.epam.reportportal.service;
 import com.epam.reportportal.listeners.ListenerParameters;
 import com.epam.reportportal.message.TypeAwareByteSource;
 import com.epam.reportportal.service.logs.LogBatchingFlowable;
+import com.epam.reportportal.service.logs.LoggingSubscriber;
 import com.epam.reportportal.utils.http.HttpRequestUtils;
 import com.epam.ta.reportportal.ws.model.BatchSaveOperatingRS;
 import com.epam.ta.reportportal.ws.model.log.SaveLogRQ;
 import com.google.common.io.ByteSource;
-import io.reactivex.Completable;
-import io.reactivex.Flowable;
-import io.reactivex.Maybe;
-import io.reactivex.Scheduler;
+import io.reactivex.*;
 import io.reactivex.functions.Function;
 import io.reactivex.internal.operators.flowable.FlowableFromObservable;
 import io.reactivex.plugins.RxJavaPlugins;
@@ -39,8 +37,6 @@ import java.io.IOException;
 import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static com.epam.reportportal.service.LoggingCallback.LOG_ERROR;
-import static com.epam.reportportal.utils.SubscriptionUtils.logFlowableResults;
 import static com.epam.reportportal.utils.files.ImageConverter.convert;
 import static com.epam.reportportal.utils.files.ImageConverter.isImage;
 import static java.util.Optional.ofNullable;
@@ -101,6 +97,25 @@ public class LoggingContext {
 	/**
 	 * Initializes new logging context and attaches it to current thread
 	 *
+	 * @param launchUuid        a UUID of a Launch
+	 * @param itemUuid          a Test Item UUID
+	 * @param client            Client of ReportPortal
+	 * @param scheduler         a {@link Scheduler} to use with this LoggingContext
+	 * @param parameters        Report Portal client configuration parameters
+	 * @param loggingSubscriber RxJava subscriber on logging results
+	 * @return New Logging Context
+	 */
+	@Nonnull
+	public static LoggingContext init(Maybe<String> launchUuid, Maybe<String> itemUuid, final ReportPortalClient client,
+			Scheduler scheduler, ListenerParameters parameters, FlowableSubscriber<BatchSaveOperatingRS> loggingSubscriber) {
+		LoggingContext context = new LoggingContext(launchUuid, itemUuid, client, scheduler, parameters, loggingSubscriber);
+		createContext().push(context);
+		return context;
+	}
+
+	/**
+	 * Initializes new logging context and attaches it to current thread
+	 *
 	 * @param launchUuid a UUID of a Launch
 	 * @param itemUuid   a Test Item UUID
 	 * @param client     Client of ReportPortal
@@ -111,9 +126,7 @@ public class LoggingContext {
 	@Nonnull
 	public static LoggingContext init(Maybe<String> launchUuid, Maybe<String> itemUuid, final ReportPortalClient client,
 			Scheduler scheduler, ListenerParameters parameters) {
-		LoggingContext context = new LoggingContext(launchUuid, itemUuid, client, scheduler, parameters);
-		createContext().push(context);
-		return context;
+		return init(launchUuid, itemUuid, client, scheduler, parameters, new LoggingSubscriber());
 	}
 
 	/**
@@ -177,7 +190,7 @@ public class LoggingContext {
 	private final boolean convertImages;
 
 	LoggingContext(Maybe<String> launchUuid, Maybe<String> itemUuid, final ReportPortalClient client, Scheduler scheduler,
-			ListenerParameters parameters) {
+			ListenerParameters parameters, FlowableSubscriber<BatchSaveOperatingRS> loggingSubscriber) {
 		this.launchUuid = launchUuid;
 		this.itemUuid = itemUuid;
 		this.emitter = PublishSubject.create();
@@ -189,10 +202,9 @@ public class LoggingContext {
 				))
 				.flatMap((Function<List<SaveLogRQ>, Flowable<BatchSaveOperatingRS>>) rqs -> client.log(HttpRequestUtils.buildLogMultiPartRequest(
 						rqs)).toFlowable())
-				.doOnError(LOG_ERROR)
 				.observeOn(scheduler)
 				.onBackpressureBuffer(parameters.getRxBufferSize(), false, true)
-				.subscribe(logFlowableResults("Logging context"));
+				.subscribe(loggingSubscriber);
 	}
 
 	private SaveLogRQ prepareRequest(String launchId, String itemId, final java.util.function.Function<String, SaveLogRQ> logSupplier)