+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - `rp.batch.payload.limit` configuration parameter, by @HardNorth
 ### Removed
 - `finally` keyword, see [JEP 421](https://openjdk.java.net/jeps/421), by @HardNorth
+### Changed
+- Logging RxJava flow refactoring, by @HardNorth
 
 ## [5.1.10]
 ### Added